+++ packages/emitter-support/src/index.ts
@@ -7,3 +7,4 @@
 export * from "./registry";
 export * from "./lowering";
 export * from "./mapping";
+export * from "./exposure";